Comment out HVG method dialog and update table export
The HVG method selection dialog in callback_DVGene2Groups.m is commented out and the default method is set to 'Splinefit Method'. The table export callback is also commented out and replaced with a call to TableViewerApp. Additionally, TableViewerApp.m has a block of code for table printing commented out.